requirements.txt
asgiref==3.4.1
Django==3.2.16
django-foundation-filefield-widget==0.0.4
packaging==23.0
Pillow==8.4.0
pytesseract==0.3.10
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.1.1
